model arch + prenet/postnet
